+++ apps/platform/features/onboarding/onboarding.hooks.ts
@@ -1,9 +1,8 @@
-import { useCallback, useEffect, useRef, useState } from 'react'
+import { useCallback, useEffect, useState } from 'react'
 import { useSelector } from 'react-redux'
 import { useRouter } from '@tanstack/react-router'
 import { useSetAtom } from 'jotai'
 
-import { useLocalStorage } from '@globalfishingwatch/react-hooks'
 import { DEFAULT_WORKSPACE_CATEGORY, DEFAULT_WORKSPACE_ID } from '@platform/config/map/workspaces'
 
 import {
@@ -13,10 +12,6 @@ import {
 } from 'features/_map/content-panel/chat/chat.atoms'
 import { useSidePanel } from 'features/_map/content-panel/contentPanel.hooks'
 import { setMapSearchOpenRequested } from 'features/_map/map/controls/map-controls.slice'
-import {
-  selectReadOnly,
-  selectScreenshotMode,
-} from 'features/_map/workspace/selectors/app.selectors'
 import { selectWorkspace } from 'features/_map/workspace/workspace.selectors'
 import { TrackCategory, trackEvent } from 'features/app/analytics.hooks'
 import { useAppDispatch } from 'features/app/app.hooks'
@@ -24,65 +19,13 @@ import type { UserGuideSlug } from 'features/cms/loaders/user-guide.types'
 import { findSectionForSlug } from 'features/help/userGuide.utils'
 import { setModalOpen } from 'features/modals/modals.slice'
 import type { OnboardingCardId } from 'features/onboarding/onboarding.config'
-import { useIsClientHydrated } from 'hooks/ssr.hooks'
-import { useAppSearch } from 'router/routes.hook'
-import { selectIsWorkspaceLocation, selectWorkspaceId } from 'router/routes.selectors'
 import { ROUTE_PATHS } from 'router/routes.utils'
 import type { QueryParams } from 'types'
-
-/** Deliberately not the legacy `WelcomePopup` key — that one may already be set from older builds. */
-export const ONBOARDING_DISMISSED_KEY = 'OnboardingPanelDismissed'
-
-export function useOnboardingDismissed() {
-  return useLocalStorage<boolean>(ONBOARDING_DISMISSED_KEY, false)
-}
 
 /** Which user guide article a slug points at, in `openSidePanel` shape. */
 function getGuideTarget(slug: UserGuideSlug) {
   const match = findSectionForSlug(slug)
   return { id: match?.section, subcontentId: match?.subSection }
-}
-
-export function useOnboardingAutoOpen() {
-  const dispatch = useAppDispatch()
-  const [dismissed] = useOnboardingDismissed()
-  const isClientHydrated = useIsClientHydrated()
-  const isWorkspaceLocation = useSelector(selectIsWorkspaceLocation)
-  const readOnly = useSelector(selectReadOnly)
-  const screenshotMode = useSelector(selectScreenshotMode)
-  const workspaceId = useSelector(selectWorkspaceId)
-  const isDefaultWorkspace = !workspaceId || workspaceId === DEFAULT_WORKSPACE_ID
-  const { sidePanelContent, dataviewInstances, dataviewInstancesOrder } = useAppSearch()
-  const hasDataviewInstances = Boolean(dataviewInstances?.length || dataviewInstancesOrder?.length)
-  const autoOpened = useRef(false)
-
-  useEffect(() => {
-    if (
-      autoOpened.current ||
-      !isClientHydrated ||
-      dismissed ||
-      readOnly ||
-      screenshotMode ||
-      !isWorkspaceLocation ||
-      !isDefaultWorkspace ||
-      hasDataviewInstances ||
-      sidePanelContent
-    ) {
-      return
-    }
-    autoOpened.current = true
-    dispatch(setModalOpen({ id: 'onboarding', open: true }))
-  }, [
-    dispatch,
-    isClientHydrated,
-    dismissed,
-    readOnly,
-    screenshotMode,
-    isWorkspaceLocation,
-    isDefaultWorkspace,
-    hasDataviewInstances,
-    sidePanelContent,
-  ])
 }
 
 /**
@@ -99,31 +42,36 @@ export function useOnboardingCardActions() {
     trackEvent({ category: TrackCategory.HelpHints, action: `onboarding panel - ${action}` })
   }, [])
 
+  const navigateWithGuide = useCallback(
+    (to: string, slug: UserGuideSlug) => {
+      const { id, subcontentId } = getGuideTarget(slug)
+      router.navigate({
+        to,
+        params: {
+          category: workspace?.category || DEFAULT_WORKSPACE_CATEGORY,
+          workspaceId: workspace?.id || DEFAULT_WORKSPACE_ID,
+        },
+        search: (prev: QueryParams): QueryParams => ({
+          ...prev,
+          sidePanelContent: 'userGuide',
+          sidePanelId: id,
+          sidePanelSubcontentId: subcontentId,
+        }),
+      })
+    },
+    [router, workspace]
+  )
+
   const onSearchVesselClick = useCallback(() => {
     track('search for a vessel')
-    // One navigation, not navigate() + openSidePanel(): the second would read a search object the
-    // first has not committed yet.
-    const { id, subcontentId } = getGuideTarget('vessel-search')
-    router.navigate({
-      to: ROUTE_PATHS.WORKSPACE_SEARCH,
-      params: {
-        category: workspace?.category || DEFAULT_WORKSPACE_CATEGORY,
-        workspaceId: workspace?.id || DEFAULT_WORKSPACE_ID,
-      },
-      search: (prev: QueryParams): QueryParams => ({
-        ...prev,
-        sidePanelContent: 'userGuide',
-        sidePanelId: id,
-        sidePanelSubcontentId: subcontentId,
-      }),
-    })
-  }, [router, workspace, track])
+    navigateWithGuide(ROUTE_PATHS.WORKSPACE_SEARCH, 'vessel-search')
+  }, [navigateWithGuide, track])
 
   const onAreaReportClick = useCallback(() => {
     track('run a report on an area')
     dispatch(setMapSearchOpenRequested(true))
-    openSidePanel({ type: 'userGuide', ...getGuideTarget('analysis-and-dynamic-reports') })
-  }, [dispatch, openSidePanel, track])
+    navigateWithGuide(ROUTE_PATHS.WORKSPACE, 'analysis-and-dynamic-reports')
+  }, [dispatch, navigateWithGuide, track])
 
   const onUserGuideClick = useCallback(() => {
     track('learn how to use the tools')
@@ -168,8 +116,6 @@ export function useOnboardingCopilotPrompt() {
         action: 'onboarding panel - ask the analysis copilot',
       })
       dispatch(setModalOpen({ id: 'onboarding', open: false }))
-      // The question starts its own conversation instead of landing in whichever thread was last
-      // open.
       setActiveThread(newActiveThread())
       setPendingPrompt(trimmed)
       openSidePanel({ type: 'chat' })
